#10 Split ClickService — extract analytics into ClickAnalyticsService
getClicksStats, getClickSummary, getFilters moved to new ClickAnalyticsService.
Corresponding actions updated to depend on ClickAnalyticsService.
ClickService retains track() and getAllClicks() (tracking + listing).
